perf: clean code
